Just verified that the MOC matches Boonth's standalone version.
This test was done in serial without using partial cell bottoms.
I also refactored a bunch to make it clearer what ghost levels
are being used where. I added in a second output from the reader
for MHT and this isn't working yet though and is causing the tests
to fail.