Fixing issue with PYTHONPATH when ParaView is run on its own.
When using UV-CDAT, we don't want to set PYTHONPATH automatically
since it also gets set in the UV-CDAT executable but it gets set
to the UV-CDAT install location which is different than the ParaView
build ones that are normally the locations that ParaView wants to set
automatically when running paraview, pvpython or pvbatch.
This fix should allow paraview to be properly used in both applications
without any problems or special environment settings.