Fixing issue of having to set PYTHONHOME for stand-alone ParaView builds.
When building ParaView with UVCDAT, we don't want to set PYTHONHOME
to the build directory location of ParaView since UVCDAT has a script
to do this and it uses the install location which causes problems.
I added an option to do this with ParaView which is on by default
and should be explicitly turned off in the UVCDAT superbuild.