Fixing issue with generated spatio-temporal script when input isn't time dependent.
Before we assumed that there was at least a single time step. For an input
that doesn't have any temporal information there wouldn't be anything outputted.
This changes it so that there will be output either way even though it's
rather silly to try spatio-temporal parallelism without any temporal
inputs.