Fixing the vtkMHTReader.
Now vtkAbstractPOPReader checks to see if the file names listed
in the meta file are also in the proper location with respect
to the meta file. This means that the full file paths don't need
to be hardcoded into the meta file anymore. I also updated the
MHT reader test to work with Boonth's test data that is now
available in UVCDAT_TestData directory.